Don't set the child instance to null when the close event fires
When re-running a process without explicitly stopping it first, the close event of the
old process is being fired after the new process is already spawned. 
This leads to a corner case where the @child variable for the new process is set to none
immediately after the close event fires for the old process.
This leads to the @stop function not working because it checks for @child.