Added -webkit-user-select
I have added 
```
.atom-runner {
    -webkit-user-select: none;
}
.atom-runner pre {
    -webkit-user-select: initial;
}
```
which first disables option to select anything within the package area. Then for the pre area it enables the option to select the generated output.

This allows you to generate output and copy and paste it somewhere.

Ideally you should instead add a button to copy the output to clipboard.